Resolved problem with $Features due to Contains method being case sensitive
The Contains method that is used throughout the Set-TargetResource
function is case sensitive. This caused a configuration to fail if the
features were specified in a case other than upper case which is how
they are specified in the function. The easiest way to resolve this
issue was to add a line to convert $Features to upper case.