Addition of Additional Modules and updates to XSQLServerSetup and xSQLServerFailoverClusterSetup
Addition of xSQLServerLogin, xSQLServerDatabaseRole,
xSQLServerDatabasePermissions, xSQLServerDatabaseOwner,
xSQLDatabaseRecoveryModel.
Update to xSQLServer and xSQLServerFailoverClusterSetup to allow users
to provide credential to Source files differing from the user which
installs SQL Server. Addition of SuprressReboot and Force Reboot flags.
Change of the $sourceFolder value to not be specific to SQL 2012.
Addition of SQL Server 2014 and 2016 support for installation.