BREAKING CHANGE: Changes to xSQLServerDatabaseRole
This resolves issue #99 and issue #108.
Added $script scope to currentPath variable
Fixed style to align with the style guide
Set-method did not throw on error
Added test for the resource xSQLServerDatabaseRole
Changes to en-US strings. Resolved that backspace was not escaped. Now it shows a backspace between server and instance in messages.