Fix v6 IPs failing Test-TargetResource
This fixes Get-TargetResource to correctly format IPv4 and IPv6 addresses in Test-TargetResource.

Because v6 addresses are input without a subnet mask, any v6 addresses had '/' appended to the end of the address with an empty subnet.

DSC would fail any run after the listener was created as the desired v6 IP would be 'f000::ba7' but the test would be expecting 'f000::ba7/'